# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "apiserver demo",
        "title": "Apiserver Example API",
        "contact": {
            "name": "lkong",
            "url": "http://www.swagger.io/support",
            "email": "466701708@qq.com"
        },
        "license": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
...
        "/api/esix/ovn/LS/{id}": {
            "delete": {
                "description": "Delete a Logical switch",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Logical switch"
                ],
                "summary": "Delete Logical switch",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Logical Switch Name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ]
            }
        },
        "/api/v1/esix/ovn/LS/{id}": {
            "get": {
                "description": "Get a Logical switch",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Logical switch"
                ],
                "summary": "GET Logical switch By name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Logical Switch Name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ]
            }
        },
        "/api/v1/esix/ovn/LS/{name}": {
            "put": {
                "description": "Add new Logical switch",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Logical switch"
                ],
                "summary": "Add new Logical switch",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Logical Switch Name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ]
            }
        },
...
    },
...
}